back/requirements.txt
# Application dependencies for TheOpenMusicBox backend
# Copyright (c) 2025 Jonathan Piette
# Licensed for non-commercial use only.

# Web framework
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
pydantic>=2.0.0
python-multipart>=0.0.6

# Socket.IO for real-time communication
python-socketio>=5.9.0
aiohttp>=3.8.0

# Audio processing
mutagen>=1.47.0

# Configuration
python-dotenv>=1.0.0

# Terminal coloring
colorama>=0.4.6

# Reactive extensions
rx>=3.2.0

# YouTube downloader
yt-dlp>=2023.10.0

# Werkzeug utilities
werkzeug>=3.0.0
